requirements.txt
cffi==1.14.0
click==7.1.1
cryptography==2.9.2
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.13.0
six==1.14.0
